Shifted full reading sample data features into NucleotideSampleData class
